refactor(tests): test directories are NeovimApplication specific
This way, any future test types (other applications besides neovim) can
have their own test directories. This is a step towards supporting
them.

This also offers a performance boost - for some reason cypress was
slower in creating the test directories than the tests themselves.
