+++ integration-tests/cypress/e2e/using-ya-to-read-events/integrations.cy.ts
@@ -1,5 +1,5 @@
 import path from "path"
-import type { IntegrationTestFile } from "server/application/neovim/testEnvironmentTypes"
+import type { IntegrationTestFile } from "server/application/neovim/environment/testEnvironmentTypes"
 import { startNeovimWithYa } from "./startNeovimWithYa"
 
 describe("grug-far integration (search and replace)", () => {
