+++ integration-tests/cypress/e2e/using-ya-to-read-events/opening-files.cy.ts
@@ -1,4 +1,4 @@
-import type { IntegrationTestFile } from "server/application/neovim/testEnvironmentTypes"
+import type { IntegrationTestFile } from "server/application/neovim/environment/testEnvironmentTypes"
 import { startNeovimWithYa } from "./startNeovimWithYa"
 
 describe("opening files", () => {
