+++ integration-tests/server/application/neovim/environment/testEnvironmentTypes.ts
@@ -54,7 +54,6 @@ export type StartNeovimServerArguments = z.infer<
 
 export const startNeovimServerArguments = z.intersection(
   z.object({
-    directory: z.string(),
     terminalDimensions: z
       .object({
         cols: z.number(),
