+++ integration-tests/server/application/neovim/neovimRouter.ts
@@ -1,15 +1,20 @@
 import { observable } from "@trpc/server/observable"
 import assert from "node:assert"
-import { Neovim, type StdoutMessage } from "server/application/neovim/Neovim"
-import { startNeovimServerArguments } from "server/application/neovim/testEnvironmentTypes"
+import {
+  NeovimApplication,
+  type StdoutMessage,
+} from "server/application/neovim/NeovimApplication"
+import { startNeovimServerArguments } from "server/application/neovim/environment/testEnvironmentTypes"
 import { trpc } from "server/connection/trpc"
-import { eventEmitter, stack } from "server/server"
+import { autocleanup, eventEmitter } from "server/server"
 import z from "zod"
 import { Lazy } from "../../utilities/Lazy"
 
+// Right now only one test instance is supported at a time. In the future, we
+// might want to support multiple test instances running in parallel
 export const neovim = new Lazy(() => {
-  const instance = new Neovim()
-  stack.use(instance)
+  const instance = new NeovimApplication()
+  autocleanup.use(instance)
   return instance
 })
 
@@ -17,10 +22,12 @@ export const neovimRouter = trpc.router({
   start: trpc.procedure
     .input(startNeovimServerArguments)
     .mutation(async (startArgs) => {
-      await neovim.get().startNextAndKillCurrent(startArgs.input)
+      const dir = await neovim.get().startNextAndKillCurrent(startArgs.input)
 
       assert(neovim.get().processId() !== undefined)
       console.log(`🚀 Started Neovim instance ${neovim.get().processId()}`)
+
+      return dir
     }),
 
   onStdout: trpc.procedure.subscription(() => {
